Slug resolver path modified and renamed files(collecSlug to slugResolver)
